+++ src/geo_base.cc
@@ -11,12 +11,12 @@
 
 namespace geo
 {
-    double deg2rad(float deg)
+    double deg2rad(double deg)
     {
         return deg * M_PI / 180;
     }
 
-    double rad2deg(float rad)
+    double rad2deg(double rad)
     {
         return rad * 180 / M_PI;
     }
